Remove Newsletter namespace
