+++ app/views/newsletter_subscriptions/_form.html.erb
@@ -1,8 +1,8 @@
-<%= form_tag(subscription_path, class: 'newsletter', remote: true) do %>
+<%= form_tag(newsletter_subscription_path, class: 'newsletter', remote: true) do %>
   <%= hidden_field_tag :authenticity_token, form_authenticity_token -%>
-  <h3 class="newsletter__heading"><%= t('newsletter.heading') %></h3>
+  <h3 class="newsletter__heading"><%= t('newsletter_subscriptions.heading') %></h3>
   <ul class="list-benefits list-benefits--s">
-    <% t('newsletter.benefits').each do |benefit| %>
+    <% t('newsletter_subscriptions.benefits').each do |benefit| %>
       <li><%= benefit %></li>
     <% end %>
   </ul>
