+++ app/views/newsletter_subscriptions/show.js.erb
