+++ config/locales/cy.yml
@@ -184,12 +184,11 @@ cy:
       newer: Diweddarach
       older: Hŷn
 
-  newsletter:
+  newsletter_subscriptions:
     heading: Negeseuon e-bost yn rhoi cyngor am ddim ar arian (Saesneg)
     benefits:
       - Offer da a newyddion am arian
       - Diolch am gofrestru
       - Awgrymiadau da ar reoli arian
-    subscription:
-      success: Y cyfan am ddim ac yn ddiduedd
-      error: Rhiowch gyfeiriad e-bost dilys
+    success: Y cyfan am ddim ac yn ddiduedd
+    error: Rhiowch gyfeiriad e-bost dilys
