+++ config/locales/en.yml
@@ -182,12 +182,11 @@ en:
       newer: Newer
       older: Older
 
-  newsletter:
+  newsletter_subscriptions:
     heading: Free money advice emails
     benefits:
       - Top tips on managing money
       - Useful tools and money news
       - All free and unbiased
-    subscription:
-      success: Thank you for signing up
-      error: Please provide a valid email address
+    success: Thank you for signing up
+    error: Please provide a valid email address
